Fix multiple instances of the bot being created (#37)
The issue of multiple instances of the bot being created was due to the
fact that the shard id becomes invalidated and so on_ready is rerun.
Therefore, to fix this, on_ready had to be replaced with setup_hook as
this coroutine only runs once. Setup_hook cannot contain an infinite
while loop, therefore send_daily_question_and_update_stats scheduled
waiting system had to be replaced with discord.ext.tasks.loop. This did
remove the force update functionality, which needs to be implemented
back again in the future.