Update roles at midday UTC rather than midnight UTC (#60)
Updating roles takes has been found to take a considerable amount of
time to process. Therefore, an investigation will be needed to fix this
inefficiency.

For the moment, updating the roles at a less important time of midday
rather than midnight (daily_reset) will make the daily_reset
notifications to be sent at the appropriate time without a delay.