Wrap all commands with created defer interaction decorator (#103)
Due to the topgg_vote_required decorator that was running code for a
duration of time before the interaction was deferred, 'interaction has
failed' errors were caused.

To resolve this, a defer_interaction decorator has been created that
defers the message. It also checks if a `display_publicly` parameter is
passed to the function which will be taken into consideration, alongside
the ephemeral default state provided, when setting the ephemeral state
of the response.

This also gave me the opportunity to move the repeated validation of the
interaction variable within this new decorator which has decluttered
most commands.

Other changes:
- Handle invalid search term for /question search.
- Update some command embed texts and colours.

> Research note: once a response is deferred, the ephemeral state of the
message cannot be changed.

I've run all the commands to see if they work with different states and
parameter values, and they all do.