feat: fetch and include video solution link in `/neetcode` embed (#167)
## Description
The NeetCode solutions information is stored in a
main.[random_letters_and_numbers].js file in neetcode.io, which can be
seen in the sources tab in developer tools. However, the random mix of
letters and numbers changes depending on the version of that file, so
the solution I found is to fetch `https://neetcode.io`, read the script
tag which mentions the most up-to-date version name of the file, then
retrieve the file's data using `https://neetcode.io/main.[...].js`. Then
the data within the file needs to be parsed using regex, converted to
json.

## Changes
- Create `NeetcodeSolutions` class that stores the solutions and
corresponding logic to fetch them from neetcode.io. The solutions are
also updated weekly through the task decorator.
- Created a function to fetch the solutions, through the method
explained in the description.
- Update the embed to display the video link, alongside pattern field,
and footer containing information on which compiled group(s) the problem
is in.
- Spoilered the code if no `||` found in the code block (because if
there is one, it will mess with the formatting)

## Notes
There is no way to actually embed a youtube link within an embed, so for
now the the URL link is just written in an embed field.