fix: spoiler tag collapses when code block contains || characters (#168)
## Description
Resolves an issue where the Discord spoiler tag would collapse if the
code block contained `||` characters.

## Changes:
Check if `response_data` contains `||` characters, replace with
`|\u200b|`

- `U+200B` is a zero-width space character. This prevents Discord from
interpreting any `||` as the end of the spoiler tag.

## Note
The character will be present in the code if copied directly from the
spoiler block. This may cause issues if the character is not accounted
for before compiling. However, this trade-off is acceptable as users
shouldn't be copying code directly without reviewing it first, and
having consistent spoiler behaviour across questions is important to
improve the user experience.

---------

Co-authored-by: Kevin Roman <44275689+Kevin-Roman@users.noreply.github.com>