fix: prevent premium questions from being displayed with /problem random (#170)
Resolves issue
[#153](https://github.com/CodeGrind-Team/CodeGrind-Bot/issues/153).

## Description
Improves the random problem retrieval process by excluding premium
questions. This is achieved by adding the `isPaidOnly` field to the
**fetch_random_question** GraphQL query. If the question is
premium-only, we simply fetch another question.

## Changes
- Modified the GraphQL query in `./utils/problems.py -
fetch_random_question` to include the `isPaidOnly` field.
- While `isPaidOnly` is True, we keep calling the
**fetch_random_question** function until a non-premium question is
found.

## Notes
This is _most likely_ the best solution. The LeetCode API only supports
filtering to display premium questions, with no option to only display
non-premium questions.

By including the `isPaidOnly` field within the **fetch_random_question**
query, we check the premium status of a question without any extra API
calls.

There are ~652 premium questions out of a possible 3199. Therefore,
while there is ~20% chance to have to retry, it is much quicker (and
provides a better user experience) to do this for them in the background
instead of making them rerun the `/problem random` command (which would
result in another request anyway).