refactor: generalise User Preferences to User Profiles and add win count support (#172)
Prerequisite for
https://github.com/CodeGrind-Team/CodeGrind-Bot/issues/160

## Description
Preference is a very specific collection, which is indexed on the server
id and user id, but it is suitable to become a more generalised
collection that could store more information about the user (such as for
the upcoming win count). So `Preference` needs to turned into a
`BaseModel` and a new document `Profile` created that contains the
preference.

## Changes
1. Rename 'Preference' document to 'Profile'. Create a new 'Preference'
basemodel that is a field within the profile documents. Update the whole
codebase to reflect these changes.
2. Create migration for converting preferences to profiles.
3. Create `update_wins` that updates all users' wins.
4. Miscellaneous function name changes.

## Post review
- [x] Migrate prod db.