fix: localize datetimes to display the correct timestamp (#177)
## The Issue
Datetimes displayed in the footer of leaderboard embeds were incorrect
(1 hour in the past).

## What has been done
The issue was caused due to the unaware datetimes that were converted to
timestamps. In order to display the correct timestamp, the datetimes had
to be converted to timezone aware datetimes.

Also, the footer question difficulty helper text has been shortened to
prevent wrapping on mobile devices.