fix(utils/schedules): create task exception wrapper
Created decorator for discord tasks such that the exceptions are handled
gracefully, and allows the task to continue to operate in it's next
schedule iteration.

Also, remove unnecessary repeated print statement of `response.status`
which may also have been leading to some exception due to it not being
wrapped in an f-string.