feat(cogs/problems): get random problem from NeetCode problem lists (#218)
## Description
Add an option to select a random question from a specific problem list.
For the scope of this PR, this includes only problem sets from NeetCode
directly.

Also added to the problem embed footer which problem lists the problem
is part of.

Space for generalisation has been made in the code to allow future
implementation of LeetCode problem sets.