fix(utils/stats): replace `html2image` with `playwright` for taking screenshots
html2image depends on a chromium based browser, and due to older/archived
chromium arm64 versions not being available, it is not possible
to continue using `html2image`.

The reason we need to use an older version of chromium is because
the new headless mode currently contains supposed bugs that lead to
issues when taking screenshots.