chore(utils/stats): replace `html2image` with `playwright` (#219)
## Description

`html2image` depends on a Chromium-based browser, and due to
older/archived Chromium arm64 versions not being available, it is not
possible to continue using `html2image`.

The reason we need to use an older version of Chromium is that the new
headless mode currently contains supposed bugs that lead to issues when
taking screenshots.

`Playwright` was chosen as it can install and setup its own Chromium
instance in headless mode, which seems not to have the issues mentioned
previously.